requirements.txt
flask
langchain
langchain-google-genai
chromadb
python-dotenv